fix(share,debug): PR-review hardening — v0.20.2
Share-from-Nextcloud (from silent-failure + code review):
- listNextcloudFolder: propagate HTTP status + error string so the picker
  can distinguish offline / 401 / 403 / 404 / 503 / captive-portal HTML
  instead of collapsing everything to "Are you offline?"
- PROPFIND now rejects non-207 responses (a captive portal's 200+HTML was
  parsing as "empty folder") and checks QXmlStreamReader::hasError() after
  the loop (malformed XML was silently truncating the list).
- listNextcloudFolder guards against empty user/server.
- shareNextcloudFileToChat: new signature takes QObject *context for
  auto-disconnect (matches fetchMentions/fetchFileImage pattern),
  removed the silent no-op default callback trap, and now passes the
  server's OCS meta.message to the caller.
- Composer's share error dialog branches on status and surfaces the
  server message; no more "HTTP 0" for network failures or "HTTP 400"
  swallowing the actual reason.

--debug flag (C3):
- Check freopen(stderr) return. On failure (AppData unreachable, perms,
  AV lock) show a Windows MessageBox explaining why no log exists,
  instead of silently defeating --debug and leaving the user confused.

Misc:
- NcFileEntry.fileId: int → qint64 (NC fileids exceed 2^31 on busy
  instances).
- Dropped dead NextcloudFilePickerDialog::m_breadcrumb field.
- Named constant for composer bars-reserve height.
- Trimmed WHAT-comments that restated code.

+++ src/core/NcFileEntry.h
@@ -4,11 +4,11 @@
 #include <QDateTime>
 
 struct NcFileEntry {
-    QString   name;           // display name ("report.pdf")
+    QString   name;
     QString   path;           // absolute, relative to user root ("/Documents/report.pdf")
     bool      isDir = false;
     qint64    size = 0;
     QDateTime lastModified;
     QString   mimeType;
-    int       fileId = 0;
+    qint64    fileId = 0;     // NC fileids routinely exceed 2^31 on busy instances
 };
